requirements.txt
blobfile==2.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
idna==3.4
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lit==15.0.7
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpi4py==3.1.6
mpmath==1.3.0
ninja==1.11.1.1
numpy==1.24.4
packaging==24.0
pillow==10.3.0
pip==24.0
piq==0.8.0
pycryptodomex==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.3.2
scipy==1.10.1
setuptools==69.5.1
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
triton==2.0.0
typing_extensions==4.11.0
urllib3==1.26.13
wheel==0.43.0
zipp==3.18.1
